Improve Portfolio Feature component
